.history/requirements_20250716164202.txt
fastapi==0.115.0
uvicorn==0.31.0
openai==1.51.0
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1